move header out to own component
